requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# NLP and Document Processing
langchain==0.1.0
langchain-openai==0.0.5
openai==1.10.0
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0

# Async processing
celery==5.3.6
redis==5.0.1

# Scheduling
APScheduler==3.10.4

# Frontend
jinja2==3.1.3
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
email-validator==2.1.0